Fractional Knapsack.md
This Java code implements the Fractional Knapsack algorithm to solve the classic knapsack problem. The problem is defined as follows: Given a set of items, each with a weight and a value, determine the maximum value that can be obtained by selecting a subset of the items such that the total weight does not exceed a given limit.